Moved utility functions out to a utils directory
